Reorganize HD wallet code and add derivation path point struct
